+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/ModuleTechBasicConfig.java
@@ -72,7 +72,7 @@ public class ModuleTechBasicConfig {
         "",
         "The whitelist takes priority over the blacklist.",
         "",
-        "String format is a recipe resource location: (domain):(path)"
+        "String format is a recipe resource location: (domain):(path) or (domain):*"
     })
     public String[] RECIPE_WHITELIST = new String[0];
 
@@ -84,7 +84,7 @@ public class ModuleTechBasicConfig {
         "",
         "The whitelist takes priority over the blacklist.",
         "",
-        "String format is a recipe resource location: (domain):(path)"
+        "String format is a recipe resource location: (domain):(path) or (domain):*"
     })
     public String[] RECIPE_BLACKLIST = new String[0];